Implementing code to ensure jwt can be used in other routes - Implementing code to have a jwt generated (for protected login purposes)

+++ backend/index.js
@@ -28,6 +28,15 @@ app.use(
   express.static(path.join(__dirname, "../client/dist/assets"))
 );
 
+const isLoggedIn = async (req, res, next) => {
+  try {
+    req.user = await findUserByToken(req.headers.authorization);
+    next();
+  } catch (ex) {
+    next(ex);
+  }
+};
+
 app.post("/api/auth/login", async (req, res, next) => {
   try {
     res.send(await authenticate(req.body));
@@ -36,9 +45,9 @@ app.post("/api/auth/login", async (req, res, next) => {
   }
 });
 
-app.get("/api/auth/me", async (req, res, next) => {
+app.get("/api/auth/me", isLoggedIn, (req, res, next) => {
   try {
-    res.send(await findUserByToken(req.headers.authorization));
+    res.send(req.user);
   } catch (ex) {
     next(ex);
   }
@@ -60,25 +69,44 @@ app.get("/api/users", async (req, res, next) => {
   }
 });
 
-app.get("/api/users/:id/userSkills", async (req, res, next) => {
+app.get("/api/users/:id/userSkills", isLoggedIn, async (req, res, next) => {
   try {
+    if (req.params.id !== req.user.id) {
+      const error = Error("not authoriized");
+      error.status = 401;
+      throw error;
+    }
     res.send(await fetchUserSkills(req.params.id));
   } catch (ex) {
     next(ex);
   }
 });
 
-app.delete("/api/users/:userId/userSkills/:id", async (req, res, next) => {
-  try {
-    await deleteUserSkill({ user_id: req.params.userId, id: req.params.id });
-    res.sendStatus(204);
-  } catch (ex) {
-    next(ex);
+app.delete(
+  "/api/users/:userId/userSkills/:id",
+  isLoggedIn,
+  async (req, res, next) => {
+    try {
+      if (req.params.id !== req.user.id) {
+        const error = Error("not authoriized");
+        error.status = 401;
+        throw error;
+      }
+      await deleteUserSkill({ user_id: req.params.userId, id: req.params.id });
+      res.sendStatus(204);
+    } catch (ex) {
+      next(ex);
+    }
   }
-});
+);
 
-app.post("/api/users/:id/userSkills", async (req, res, next) => {
+app.post("/api/users/:id/userSkills", isLoggedIn, async (req, res, next) => {
   try {
+    if (req.params.id !== req.user.id) {
+      const error = Error("not authoriized");
+      error.status = 401;
+      throw error;
+    }
     res.status(201).send(
       await createUserSkill({
         user_id: req.params.id,
